# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Awesome API",
        "description": "This is the API documentation for My Awesome Project.",
        "contact": {
            "email": "support@myawesomeapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/usersList": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "List Of Users",
                "description": "Authenticate user with username and password, and return data if role is 1",
                "operationId": "2c2f8b46a278779b3ed61eb217f21dfe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User role is 1",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/create-task": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create New Task",
                "description": "Submit new task for your self or anyone have been register in system",
                "operationId": "46ddacb92ed38d1812a00f0e795f0a90",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password",
                                    "task_name",
                                    "description",
                                    "due_date",
                                    "assigned_to"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "task_name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "due_date": {
                                        "type": "string"
                                    },
                                    "assigned_to": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/list-task": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "List Of Task",
                "description": "Response List Of Tasks",
                "operationId": "f5adab24ddee2f4035a5df9a55a95a73",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "tasks": {
                                            "$ref": "#/components/schemas/Tasks"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/update-status-task": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update An Existing Task",
                "description": "Update an exsiting task for your self or anyone have been register in system",
                "operationId": "ae651d7cfe8318b2cec7ec141bd804e2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password",
                                    "task_id",
                                    "task_name",
                                    "description",
                                    "due_date",
                                    "assigned_to",
                                    "status"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "task_id": {
                                        "type": "string"
                                    },
                                    "task_name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "due_date": {
                                        "type": "string"
                                    },
                                    "assigned_to": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Task Not Found"
                    }
                }
            }
        },
        "/api/v1/delete-task": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete An Existing Task",
                "description": "Delete an exsiting task for your self or anyone have been register in system when not role is admin",
                "operationId": "b368c79f14e9c144bcdca6970ff30469",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password",
                                    "task_id"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "task_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Task Not Found"
                    }
                }
            }
        },
        "/api/v1/counter-task": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Counter Of Task",
                "description": "Response Counter Of Tasks",
                "operationId": "01c8b4d27d2302edf3935b2a9dcc7c8d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "taskall": {
                                            "type": "string",
                                            "example": "0"
                                        },
                                        "openTaskCount": {
                                            "type": "string",
                                            "example": "0"
                                        },
                                        "closeTaskCount": {
                                            "type": "string",
                                            "example": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "User's name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User's email address",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Tasks": {
                "required": [
                    "id",
                    "task_name",
                    "description",
                    "assigned_to",
                    "due_date",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "task_name": {
                        "description": "Task's name",
                        "type": "string"
                    },
                    "description": {
                        "description": "Task's description",
                        "type": "string"
                    },
                    "assigned_to": {
                        "description": "Task's assigned_to",
                        "type": "integer"
                    },
                    "due_date": {
                        "description": "Task's due_date",
                        "type": "string"
                    },
                    "status": {
                        "description": "Task's status",
                        "type": "boolean"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Tasks",
            "description": "Tasks"
        }
    ]
}